Add session facility and global navigation button function
Sessions have now been enabled in the BaseController, thus enabling them in all other controllers that are extending the class.

The global button command will allow the return of different sets of buttons depending on factors such as if the user is logged in, as well as if they are an administrator.

At present, different buttons are returned if the user is logged in (though the login functionality is not done yet)