Add login controller & logout functionality (LOGIN NOT STARTED YET)
The login controller will allow the return of pages to facilitate logging in and signing up, but this isn't done yet.

A method to remove the user's session, thus logging them out, is written with a redirection to the home page. This is written in advance as the database isn't fully functional yet.